+++ config/hashtagcmscommon.php
@@ -1,5 +1,5 @@
 <?php
 
 return [
-    'version'=>'1.3.8'
+    'version'=>'1.4.0'
 ];